Refactor; extract ContentDisposition header functions.
